+++ widgets/sparkline/sparklinedemo/sparklinedemo.go
@@ -51,6 +51,30 @@ func playSparkLine(ctx context.Context, sl *sparkline.SparkLine, delay time.Dura
 	}
 }
 
+// fillSparkLine continuously fills the SparkLine up to its capacity with
+// random values.
+func fillSparkLine(ctx context.Context, sl *sparkline.SparkLine, delay time.Duration) {
+	const max = 100
+
+	ticker := time.NewTicker(delay)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-ticker.C:
+			var values []int
+			for i := 0; i < sl.ValueCapacity(); i++ {
+				values = append(values, int(rand.Int31n(max+1)))
+			}
+			if err := sl.Add(values); err != nil {
+				panic(err)
+			}
+
+		case <-ctx.Done():
+			return
+		}
+	}
+}
+
 func main() {
 	t, err := termbox.New()
 	if err != nil {
@@ -82,7 +106,7 @@ func main() {
 	if err != nil {
 		panic(err)
 	}
-	go playSparkLine(ctx, yellow, 1*time.Second)
+	go fillSparkLine(ctx, yellow, 1*time.Second)
 
 	c, err := container.New(
 		t,